refactor(backend): Port FileUtils to FileService
Ports the FileUtils to the injectable FileService application scoped
bean.
